Improve UX: 設定画面を階層化してユーザビリティを向上
## 改善内容

### 🏗️ 設定画面の階層化
- MainSettingsScreen: メイン設定画面（カテゴリ一覧）
- GitSettingsScreen: データ同期設定専用画面
- NotificationSettingsScreen: 通知設定専用画面
- AboutScreen: アプリ情報画面

### 📱 UX改善
- 関心の分離: 各設定が独立して管理可能
- 認知負荷の軽減: 一度に表示される情報量を適切に調整
- 拡張性: 新しい設定カテゴリを追加しやすい構造
- 一般的なパターン: 多くのアプリで採用されている階層構造

### 🎨 UI改善
- カテゴリ別のアイコン表示
- 各設定項目の説明文追加
- 統一されたカードデザイン
- 適切な余白とレイアウト

### 🔧 技術的改善
- ナビゲーション構造の最適化
- コンポーネントの再利用性向上
- 保守性の向上

これにより、ユーザーが目的の設定を素早く見つけられ、
設定変更時の迷いを減らすことができます。
